requirements.txt
pandas
deepl
python-docx
openpyxl
ttkbootstrap
tk